reflective_vision_system/requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0

# Detection
ultralytics>=8.0.0

# Segmentation
segment-anything>=1.0

# Depth Estimation
transformers>=4.35.0
timm>=0.9.0

# Tracking
filterpy>=1.4.5

# Data & Config
PyYAML>=6.0
albumentations>=1.3.0
pycocotools>=2.0.7

# Visualization & Utils
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
rich>=13.0.0

# Experiment Tracking (optional)
tensorboard>=2.14.0
